Unity Atoms Input System Integration
This is just a quick demo to suggest to the Unity Atoms team. There 'was' a line on the github that said to contact you... but creating a Pull Request is the more funner way of doing things.

Here we have a YouTube link that explains what these scripts are for. The scene is a bit bare bones since testbed and all that.
https://www.youtube.com/watch?v=q7W8FyTIriQ&feature=youtu.be