Refactor VariableDrawer for generic serialization
VariableDrawer is now a drawer for the generic type AtomBaseVariable<>. To be useable it was necesarry to remove its abtract and generic modifier, which in turn meant AtomDrawer could no longer be generic, which in turn meant that every drawer that inherits from AtomDrawer had to be deleted. Ultimately, this is a good thing, but I had not expected the purge to come quite so soon.