Make AtomDrawer's Create button generic
The create button is a very important aspect of Unity Atoms and this upgrade makes it so that it can be used by generic instances as well. Once it is clicked on a generic instance, all inherited types are found and the user can select what type he wants to create.
Currently this only works for types that inherit from AtomBaseVariable<>, but you can expect this same behaviour for Events and Lists in later commits.