Focus Text when AtomDrawer Create is clicked
This is a very comfortable quality of life improvement, where clicking create will auto-focus the textfield for naming the atom. Unfortunately I had to jump through a hoop to get it to work for Create with multiple options (trust me, I've tried to make it easier than what it currently is.) For now though this work perfectly and feels very nice working with in the Editor.