Add Object Picker for generic types
Generic fields have an object picker that doesn't work. This commit goes in a roundabout way to do it, but it is glorious and works beautifully. The things I've tried to get this to work... anyway.

A button is drawn on top of the object picker with no styling. This gives the illusion that you are pressing the regular 'ol object picker, while in actually this button opens our own object picker. This is then given a filter to look for all types that inherit from the generic class. A dictionary is used to couple the selected object back to the property (possibly this can be integrated in the dictionary that was already there).