Add Safe Search toggle to AtomGenerator
Previously, the AtomGenerator was only able to generate types that were Unity compatible and either Serializable or ScriptableObjects.

The first change is that it now also searches for Objects to account for prefabs.
The second change is a toggle next to the generator to differentiate between searching for Unity compatible types and searching for all types (denoted as "unsafe types"). This does what it promises: it gets all the (dynamic) assemblies it can find, retrieves all exported types, and makes no distinction between serializable or nonserializable. This returns the functionality to generate for example strings and ints with the generator.