Generic property drawers (#226)
* Remove redundant #if directive

* Refactor VariableDrawer for generic serialization

VariableDrawer is now a drawer for the generic type AtomBaseVariable<>. To be useable it was necesarry to remove its abtract and generic modifier, which in turn meant AtomDrawer could no longer be generic, which in turn meant that every drawer that inherits from AtomDrawer had to be deleted. Ultimately, this is a good thing, but I had not expected the purge to come quite so soon.

* Make AtomDrawer's Create button generic

The create button is a very important aspect of Unity Atoms and this upgrade makes it so that it can be used by generic instances as well. Once it is clicked on a generic instance, all inherited types are found and the user can select what type he wants to create.
Currently this only works for types that inherit from AtomBaseVariable<>, but you can expect this same behaviour for Events and Lists in later commits.

* Make AtomDrawer an inherited drawer

Anything inheriting from BaseAtom is now drawn without the need for additional drawers, including but not limited to Events and ValueLists.

* Fix duplicate named asset creation

AtomDrawer could create an asset at a path that already existed, however it will now create a unique asset path for the new asset.

* Focus Text when AtomDrawer Create is clicked

This is a very comfortable quality of life improvement, where clicking create will auto-focus the textfield for naming the atom. Unfortunately I had to jump through a hoop to get it to work for Create with multiple options (trust me, I've tried to make it easier than what it currently is.) For now though this work perfectly and feels very nice working with in the Editor.

* Focus Text when AtomDrawer Create is clicked

This is a very comfortable quality of life improvement, where clicking create will auto-focus the textfield for naming the atom. Unfortunately I had to jump through a hoop to get it to work for Create with multiple options (trust me, I've tried to make it easier than what it currently is.) For now though this work perfectly and feels very nice working with in the Editor.

* Add an optimization to the Create button

If the type is not generic and sealed, then don't search the assemblies. This probably doesn't matter at all, but it was simple enough to add.

* Add Object Picker for generic types

Generic fields have an object picker that doesn't work. This commit goes in a roundabout way to do it, but it is glorious and works beautifully. The things I've tried to get this to work... anyway.

A button is drawn on top of the object picker with no styling. This gives the illusion that you are pressing the regular 'ol object picker, while in actually this button opens our own object picker. This is then given a filter to look for all types that inherit from the generic class. A dictionary is used to couple the selected object back to the property (possibly this can be integrated in the dictionary that was already there).

* Use GUIUtility to get an object picker controlID

* Fix AtomDrawer PrefixLabel showing up in Reference<T>

It will also remove a redundant line from the documentation (V5 will always at least be version 2020.1)

* Remove AtomDrawers from the Generator

* Fix the Generator

Note that the AtomEditors are temporary and are in the process of being removed entirely.

* Fix formating in RegenerateAllAtoms

* Add auto focus naming field if creation failed

* Add keyboard shortcuts to confirm/deny create