Add improved Generator methods
The methods take a type and a resolver. The resolver is used to generate the atom, the type is used to find what values to put in.
This modular approach makes it incredibly easy and straightforward to add new resolvers, as can be seen in the Generator class on line 208, where Template.ResolveAtom is given as a default value.